wrapper class handling user login
