+++ src/frontend/lib/services/auth_service.dart
@@ -1,0 +1,14 @@
+import 'package:firebase_auth/firebase_auth.dart';
+
+/// Handles user authentication by communicating with Firebase Authentication
+class AuthService {
+
+  final FirebaseAuth _auth = FirebaseAuth.instance;
+
+  // Sign in with email and password
+
+  // Register with email and password
+
+  // Sign out
+
+}
