+++ src/frontend/lib/views/authenticate/authenticate.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class Authenticate extends StatefulWidget {
+  @override
+  _AuthenticateState createState() => _AuthenticateState();
+}
+
+class _AuthenticateState extends State<Authenticate> {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: Text('Authentication screen'),
+    );
+  }
+}
